+++ tagmanager/src/tm_tag.c
@@ -11,13 +11,7 @@
 #include <string.h>
 #include <glib-object.h>
 
-#include "general.h"
-#include "entry.h"
-#include "parse.h"
-#include "read.h"
-#define LIBCTAGS_DEFINED
 #include "tm_tag.h"
-#include "tm_parser.h"
 
 
 #define TAG_NEW(T)	((T) = g_slice_new0(TMTag))
@@ -79,32 +73,6 @@ static void log_tag_free(TMTag *tag)
 #endif /* DEBUG_TAG_REFS */
 
 
-const TMTagType TM_GLOBAL_TYPE_MASK =
-	tm_tag_class_t | tm_tag_enum_t | tm_tag_interface_t |
-	tm_tag_struct_t | tm_tag_typedef_t | tm_tag_union_t | tm_tag_namespace_t;
-
-
-/* Note: To preserve binary compatibility, it is very important
-	that you only *append* to this list ! */
-enum
-{
-	TA_NAME = 200,
-	TA_LINE,
-	TA_LOCAL,
-	TA_POS, /* Obsolete */
-	TA_TYPE,
-	TA_ARGLIST,
-	TA_SCOPE,
-	TA_VARTYPE,
-	TA_INHERITS,
-	TA_TIME,
-	TA_ACCESS,
-	TA_IMPL,
-	TA_LANG,
-	TA_INACTIVE, /* Obsolete */
-	TA_POINTER
-};
-
 typedef struct
 {
 	guint *sort_attrs;
@@ -112,48 +80,6 @@ typedef struct
 	const GPtrArray *tags_array;
 	gboolean first;
 } TMSortOptions;
-
-static const char *s_tag_type_names[] = {
-	"class", /* classes */
-	"enum", /* enumeration names */
-	"enumerator", /* enumerators (values inside an enumeration) */
-	"externvar", /* external variable declarations */
-	"field", /* fields */
-	"function", /*  function definitions */
-	"interface", /* interfaces */
-	"macro", /* macro definitions */
-	"member", /* class, struct, and union members */
-	"method", /* methods */
-	"namespace", /* namespaces */
-	"package", /* packages */
-	"prototype", /* function prototypes */
-	"struct", /* structure names */
-	"typedef", /* typedefs */
-	"union", /* union names */
-	"variable", /* variable definitions */
-	"other" /* Other tag type (non C/C++/Java) */
-};
-
-static TMTagType s_tag_types[] = {
-	tm_tag_class_t,
-	tm_tag_enum_t,
-	tm_tag_enumerator_t,
-	tm_tag_externvar_t,
-	tm_tag_field_t,
-	tm_tag_function_t,
-	tm_tag_interface_t,
-	tm_tag_macro_t,
-	tm_tag_member_t,
-	tm_tag_method_t,
-	tm_tag_namespace_t,
-	tm_tag_package_t,
-	tm_tag_prototype_t,
-	tm_tag_struct_t,
-	tm_tag_typedef_t,
-	tm_tag_union_t,
-	tm_tag_variable_t,
-	tm_tag_other_t
-};
 
 /* Gets the GType for a TMTag */
 GType tm_tag_get_type(void)
@@ -167,467 +93,18 @@ GType tm_tag_get_type(void)
 	return gtype;
 }
 
-static TMTagType get_tag_type(const char *tag_name)
-{
-	unsigned int i;
-	int cmp;
-	g_return_val_if_fail(tag_name, 0);
-	for (i=0; i < sizeof(s_tag_type_names)/sizeof(char *); ++i)
-	{
-		cmp = strcmp(tag_name, s_tag_type_names[i]);
-		if (0 == cmp)
-			return s_tag_types[i];
-		else if (cmp < 0)
-			break;
-	}
-	/* other is not checked above as it is last, not sorted alphabetically */
-	if (strcmp(tag_name, "other") == 0)
-		return tm_tag_other_t;
-#ifdef TM_DEBUG
-	fprintf(stderr, "Unknown tag type %s\n", tag_name);
-#endif
-	return tm_tag_undef_t;
-}
-
-static char get_tag_impl(const char *impl)
-{
-	if ((0 == strcmp("virtual", impl))
-	 || (0 == strcmp("pure virtual", impl)))
-		return TAG_IMPL_VIRTUAL;
-
-#ifdef TM_DEBUG
-		g_warning("Unknown implementation %s", impl);
-#endif
-	return TAG_IMPL_UNKNOWN;
-}
-
-static char get_tag_access(const char *access)
-{
-	if (0 == strcmp("public", access))
-		return TAG_ACCESS_PUBLIC;
-	else if (0 == strcmp("protected", access))
-		return TAG_ACCESS_PROTECTED;
-	else if (0 == strcmp("private", access))
-		return TAG_ACCESS_PRIVATE;
-	else if (0 == strcmp("friend", access))
-		return TAG_ACCESS_FRIEND;
-	else if (0 == strcmp("default", access))
-		return TAG_ACCESS_DEFAULT;
-
-#ifdef TM_DEBUG
-	g_warning("Unknown access type %s", access);
-#endif
-	return TAG_ACCESS_UNKNOWN;
-}
-
 /*
- Initializes a TMTag structure with information from a tagEntryInfo struct
- used by the ctags parsers. Note that the TMTag structure must be malloc()ed
- before calling this function. This function is called by tm_tag_new() - you
- should not need to call this directly.
- @param tag The TMTag structure to initialize
- @param file Pointer to a TMSourceFile struct (it is assigned to the file member)
- @param tag_entry Tag information gathered by the ctags parser
- @return TRUE on success, FALSE on failure
-*/
-static gboolean tm_tag_init(TMTag *tag, TMSourceFile *file, const tagEntryInfo *tag_entry)
-{
-	tag->refcount = 1;
-	if (NULL == tag_entry)
-		return FALSE;
-		
-	/* This is a normal tag entry */
-	if (NULL == tag_entry->name)
-		return FALSE;
-	tag->name = g_strdup(tag_entry->name);
-	tag->type = get_tag_type(tag_entry->kindName);
-	tag->local = tag_entry->isFileScope;
-	tag->pointerOrder = 0;	/* backward compatibility (use var_type instead) */
-	tag->line = tag_entry->lineNumber;
-	if (NULL != tag_entry->extensionFields.arglist)
-		tag->arglist = g_strdup(tag_entry->extensionFields.arglist);
-	if ((NULL != tag_entry->extensionFields.scope[1]) &&
-		(0 != tag_entry->extensionFields.scope[1][0]))
-		tag->scope = g_strdup(tag_entry->extensionFields.scope[1]);
-	if (tag_entry->extensionFields.inheritance != NULL)
-		tag->inheritance = g_strdup(tag_entry->extensionFields.inheritance);
-	if (tag_entry->extensionFields.varType != NULL)
-		tag->var_type = g_strdup(tag_entry->extensionFields.varType);
-	if (tag_entry->extensionFields.access != NULL)
-		tag->access = get_tag_access(tag_entry->extensionFields.access);
-	if (tag_entry->extensionFields.implementation != NULL)
-		tag->impl = get_tag_impl(tag_entry->extensionFields.implementation);
-	if ((tm_tag_macro_t == tag->type) && (NULL != tag->arglist))
-		tag->type = tm_tag_macro_with_arg_t;
-	tag->file = file;
-	tag->lang = file->lang;
-	return TRUE;
-}
-
-/*
- Creates a new tag structure from a tagEntryInfo pointer and a TMSOurceFile pointer
- and returns a pointer to it.
- @param file - Pointer to the TMSourceFile structure containing the tag
- @param tag_entry Contains tag information generated by ctags
+ Creates a new tag structure and returns a pointer to it.
  @return the new TMTag structure. This should be free()-ed using tm_tag_free()
 */
-TMTag *tm_tag_new(TMSourceFile *file, const tagEntryInfo *tag_entry)
+TMTag *tm_tag_new(void)
 {
 	TMTag *tag;
 
 	TAG_NEW(tag);
-	if (FALSE == tm_tag_init(tag, file, tag_entry))
-	{
-		TAG_FREE(tag);
-		return NULL;
-	}
+	tag->refcount = 1;
+
 	return tag;
-}
-
-/*
- Initializes an already malloc()ed TMTag structure by reading a tag entry
- line from a file. The structure should be allocated beforehand.
- @param tag The TMTag structure to populate
- @param file The TMSourceFile struct (assigned to the file member)
- @param fp FILE pointer from where the tag line is read
- @return TRUE on success, FALSE on FAILURE
-*/
-static gboolean tm_tag_init_from_file(TMTag *tag, TMSourceFile *file, FILE *fp)
-{
-	guchar buf[BUFSIZ];
-	guchar *start, *end;
-	gboolean status;
-	guchar changed_char = TA_NAME;
-
-	tag->refcount = 1;
-	if ((NULL == fgets((gchar*)buf, BUFSIZ, fp)) || ('\0' == *buf))
-		return FALSE;
-	for (start = end = buf, status = TRUE; (TRUE == status); start = end, ++ end)
-	{
-		while ((*end < TA_NAME) && (*end != '\0') && (*end != '\n'))
-			++ end;
-		if (('\0' == *end) || ('\n' == *end))
-			status = FALSE;
-		changed_char = *end;
-		*end = '\0';
-		if (NULL == tag->name)
-		{
-			if (!isprint(*start))
-				return FALSE;
-			else
-				tag->name = g_strdup((gchar*)start);
-		}
-		else
-		{
-			switch (*start)
-			{
-				case TA_LINE:
-					tag->line = atol((gchar*)start + 1);
-					break;
-				case TA_LOCAL:
-					tag->local = atoi((gchar*)start + 1);
-					break;
-				case TA_TYPE:
-					tag->type = (TMTagType) atoi((gchar*)start + 1);
-					break;
-				case TA_ARGLIST:
-					tag->arglist = g_strdup((gchar*)start + 1);
-					break;
-				case TA_SCOPE:
-					tag->scope = g_strdup((gchar*)start + 1);
-					break;
-				case TA_POINTER:
-					tag->pointerOrder = atoi((gchar*)start + 1);
-					break;
-				case TA_VARTYPE:
-					tag->var_type = g_strdup((gchar*)start + 1);
-					break;
-				case TA_INHERITS:
-					tag->inheritance = g_strdup((gchar*)start + 1);
-					break;
-				case TA_TIME:  /* Obsolete */
-					break;
-				case TA_LANG:  /* Obsolete */
-					break;
-				case TA_INACTIVE:  /* Obsolete */
-					break;
-				case TA_ACCESS:
-					tag->access = (char) *(start + 1);
-					break;
-				case TA_IMPL:
-					tag->impl = (char) *(start + 1);
-					break;
-				default:
-#ifdef GEANY_DEBUG
-					g_warning("Unknown attribute %s", start + 1);
-#endif
-					break;
-			}
-		}
-		*end = changed_char;
-	}
-	if (NULL == tag->name)
-		return FALSE;
-	tag->file = file;
-	return TRUE;
-}
-
-/* alternative parser for Pascal and LaTeX global tags files with the following format
- * tagname|return value|arglist|description\n */
-static gboolean tm_tag_init_from_file_alt(TMTag *tag, TMSourceFile *file, FILE *fp)
-{
-	guchar buf[BUFSIZ];
-	guchar *start, *end;
-	gboolean status;
-	/*guchar changed_char = TA_NAME;*/
-
-	tag->refcount = 1;
-	if ((NULL == fgets((gchar*)buf, BUFSIZ, fp)) || ('\0' == *buf))
-		return FALSE;
-	{
-		gchar **fields;
-		guint field_len;
-		for (start = end = buf, status = TRUE; (TRUE == status); start = end, ++ end)
-		{
-			while ((*end < TA_NAME) && (*end != '\0') && (*end != '\n'))
-				++ end;
-			if (('\0' == *end) || ('\n' == *end))
-				status = FALSE;
-			/*changed_char = *end;*/
-			*end = '\0';
-			if (NULL == tag->name && !isprint(*start))
-					return FALSE;
-
-			fields = g_strsplit((gchar*)start, "|", -1);
-			field_len = g_strv_length(fields);
-
-			if (field_len >= 1) tag->name = g_strdup(fields[0]);
-			else tag->name = NULL;
-			if (field_len >= 2 && fields[1] != NULL) tag->var_type = g_strdup(fields[1]);
-			if (field_len >= 3 && fields[2] != NULL) tag->arglist = g_strdup(fields[2]);
-			tag->type = tm_tag_prototype_t;
-			g_strfreev(fields);
-		}
-	}
-
-	if (NULL == tag->name)
-		return FALSE;
-	tag->file = file;
-	return TRUE;
-}
-
-/*
- Same as tm_tag_init_from_file(), but parsing CTags tag file format
- (http://ctags.sourceforge.net/FORMAT) 
-*/
-static gboolean tm_tag_init_from_file_ctags(TMTag *tag, TMSourceFile *file, FILE *fp)
-{
-	gchar buf[BUFSIZ];
-	gchar *p, *tab;
-
-	tag->refcount = 1;
-	tag->type = tm_tag_function_t; /* default type is function if no kind is specified */
-	do
-	{
-		if ((NULL == fgets(buf, BUFSIZ, fp)) || ('\0' == *buf))
-			return FALSE;
-	}
-	while (strncmp(buf, "!_TAG_", 6) == 0); /* skip !_TAG_ lines */
-
-	p = buf;
-
-	/* tag name */
-	if (! (tab = strchr(p, '\t')) || p == tab)
-		return FALSE;
-	tag->name = g_strndup(p, (gsize)(tab - p));
-	p = tab + 1;
-
-	/* tagfile, unused */
-	if (! (tab = strchr(p, '\t')))
-	{
-		g_free(tag->name);
-		tag->name = NULL;
-		return FALSE;
-	}
-	p = tab + 1;
-	/* Ex command, unused */
-	if (*p == '/' || *p == '?')
-	{
-		gchar c = *p;
-		for (++p; *p && *p != c; p++)
-		{
-			if (*p == '\\' && p[1])
-				p++;
-		}
-	}
-	else /* assume a line */
-		tag->line = atol(p);
-	tab = strstr(p, ";\"");
-	/* read extension fields */
-	if (tab)
-	{
-		p = tab + 2;
-		while (*p && *p != '\n' && *p != '\r')
-		{
-			gchar *end;
-			const gchar *key, *value = NULL;
-
-			/* skip leading tabulations */
-			while (*p && *p == '\t') p++;
-			/* find the separator (:) and end (\t) */
-			key = end = p;
-			while (*end && *end != '\t' && *end != '\n' && *end != '\r')
-			{
-				if (*end == ':' && ! value)
-				{
-					*end = 0; /* terminate the key */
-					value = end + 1;
-				}
-				end++;
-			}
-			/* move p paste the so we won't stop parsing by setting *end=0 below */
-			p = *end ? end + 1 : end;
-			*end = 0; /* terminate the value (or key if no value) */
-
-			if (! value || 0 == strcmp(key, "kind")) /* tag kind */
-			{
-				const gchar *kind = value ? value : key;
-
-				if (kind[0] && kind[1])
-					tag->type = get_tag_type(kind);
-				else
-				{
-					switch (*kind)
-					{
-						case 'c': tag->type = tm_tag_class_t; break;
-						case 'd': tag->type = tm_tag_macro_t; break;
-						case 'e': tag->type = tm_tag_enumerator_t; break;
-						case 'F': tag->type = tm_tag_other_t; break;  /* Obsolete */
-						case 'f': tag->type = tm_tag_function_t; break;
-						case 'g': tag->type = tm_tag_enum_t; break;
-						case 'I': tag->type = tm_tag_class_t; break;
-						case 'i': tag->type = tm_tag_interface_t; break;
-						case 'l': tag->type = tm_tag_variable_t; break;
-						case 'M': tag->type = tm_tag_macro_t; break;
-						case 'm': tag->type = tm_tag_member_t; break;
-						case 'n': tag->type = tm_tag_namespace_t; break;
-						case 'P': tag->type = tm_tag_package_t; break;
-						case 'p': tag->type = tm_tag_prototype_t; break;
-						case 's': tag->type = tm_tag_struct_t; break;
-						case 't': tag->type = tm_tag_typedef_t; break;
-						case 'u': tag->type = tm_tag_union_t; break;
-						case 'v': tag->type = tm_tag_variable_t; break;
-						case 'x': tag->type = tm_tag_externvar_t; break;
-						default:
-#ifdef TM_DEBUG
-							g_warning("Unknown tag kind %c", *kind);
-#endif
-							tag->type = tm_tag_other_t; break;
-					}
-				}
-			}
-			else if (0 == strcmp(key, "inherits")) /* comma-separated list of classes this class inherits from */
-			{
-				g_free(tag->inheritance);
-				tag->inheritance = g_strdup(value);
-			}
-			else if (0 == strcmp(key, "implementation")) /* implementation limit */
-				tag->impl = get_tag_impl(value);
-			else if (0 == strcmp(key, "line")) /* line */
-				tag->line = atol(value);
-			else if (0 == strcmp(key, "access")) /* access */
-				tag->access = get_tag_access(value);
-			else if (0 == strcmp(key, "class") ||
-					 0 == strcmp(key, "enum") ||
-					 0 == strcmp(key, "function") ||
-					 0 == strcmp(key, "struct") ||
-					 0 == strcmp(key, "union")) /* Name of the class/enum/function/struct/union in which this tag is a member */
-			{
-				g_free(tag->scope);
-				tag->scope = g_strdup(value);
-			}
-			else if (0 == strcmp(key, "file")) /* static (local) tag */
-				tag->local = TRUE;
-			else if (0 == strcmp(key, "signature")) /* arglist */
-			{
-				g_free(tag->arglist);
-				tag->arglist = g_strdup(value);
-			}
-		}
-	}
-
-	tag->file = file;
-	return TRUE;
-}
-
-/*
- Same as tm_tag_new() except that the tag attributes are read from file.
- @param mode TMParserType to use for the tag.
-*/
-TMTag *tm_tag_new_from_file(TMSourceFile *file, FILE *fp, TMParserType mode, TMFileFormat format)
-{
-	TMTag *tag;
-	gboolean result = FALSE;
-
-	TAG_NEW(tag);
-
-	switch (format)
-	{
-		case TM_FILE_FORMAT_TAGMANAGER:
-			result = tm_tag_init_from_file(tag, file, fp);
-			break;
-		case TM_FILE_FORMAT_PIPE:
-			result = tm_tag_init_from_file_alt(tag, file, fp);
-			break;
-		case TM_FILE_FORMAT_CTAGS:
-			result = tm_tag_init_from_file_ctags(tag, file, fp);
-			break;
-	}
-
-	if (! result)
-	{
-		TAG_FREE(tag);
-		return NULL;
-	}
-	tag->lang = mode;
-	return tag;
-}
-
-/*
- Writes tag information to the given FILE *.
- @param tag The tag information to write.
- @param file FILE pointer to which the tag information is written.
- @param attrs Attributes to be written (bitmask).
- @return TRUE on success, FALSE on failure.
-*/
-gboolean tm_tag_write(TMTag *tag, FILE *fp, TMTagAttrType attrs)
-{
-	fprintf(fp, "%s", tag->name);
-	if (attrs & tm_tag_attr_type_t)
-		fprintf(fp, "%c%d", TA_TYPE, tag->type);
-	if ((attrs & tm_tag_attr_arglist_t) && (NULL != tag->arglist))
-		fprintf(fp, "%c%s", TA_ARGLIST, tag->arglist);
-	if (attrs & tm_tag_attr_line_t)
-		fprintf(fp, "%c%ld", TA_LINE, tag->line);
-	if (attrs & tm_tag_attr_local_t)
-		fprintf(fp, "%c%d", TA_LOCAL, tag->local);
-	if ((attrs & tm_tag_attr_scope_t) && (NULL != tag->scope))
-		fprintf(fp, "%c%s", TA_SCOPE, tag->scope);
-	if ((attrs & tm_tag_attr_inheritance_t) && (NULL != tag->inheritance))
-		fprintf(fp, "%c%s", TA_INHERITS, tag->inheritance);
-	if (attrs & tm_tag_attr_pointer_t)
-		fprintf(fp, "%c%d", TA_POINTER, tag->pointerOrder);
-	if ((attrs & tm_tag_attr_vartype_t) && (NULL != tag->var_type))
-		fprintf(fp, "%c%s", TA_VARTYPE, tag->var_type);
-	if ((attrs & tm_tag_attr_access_t) && (TAG_ACCESS_UNKNOWN != tag->access))
-		fprintf(fp, "%c%c", TA_ACCESS, tag->access);
-	if ((attrs & tm_tag_attr_impl_t) && (TAG_IMPL_UNKNOWN != tag->impl))
-		fprintf(fp, "%c%c", TA_IMPL, tag->impl);
-
-	if (fprintf(fp, "\n"))
-		return TRUE;
-	else
-		return FALSE;
 }
 
 /*
